Sick if else statment!

+++ db/migrate/20210105014935_create_transactions.rb
@@ -1,8 +1,8 @@
 class CreateTransactions < ActiveRecord::Migration[5.2]
   def change
     create_table :transactions do |t|
-      t.integer :credit_card_number
-      t.integer :credit_card_expiration_date
+      t.bigint :credit_card_number
+      t.bigint :credit_card_expiration_date
       t.integer :result
       t.references :invoice, foreign_key: true
